+++ app/cache/FileCache.php
@@ -373,7 +373,6 @@ class FileCache implements CacheInterface
         } finally {
             flock($fp, LOCK_UN);
             fclose($fp);
-            @unlink($lockFile);
         }
     }
 
@@ -403,7 +402,6 @@ class FileCache implements CacheInterface
         } finally {
             flock($fp, LOCK_UN);
             fclose($fp);
-            @unlink($lockFile);
         }
     }
 
@@ -433,7 +431,6 @@ class FileCache implements CacheInterface
         } finally {
             flock($fp, LOCK_UN);
             fclose($fp);
-            @unlink($lockFile);
         }
     }